Provides unique user names
This Keypirinha package didn't work for me at all
because the field 'name' was not set for some users.
Another problem was that that the field 'name' was
not unique and they were displayed with the name.
I had no change to guess which is the one I was looking
for. That's why I added the username to the display value.